Move files.ts to /common
