+++ extensions/ql-vscode/src/common/files.ts
