+++ extensions/ql-vscode/src/query-history/query-history-scrubber.ts
@@ -3,7 +3,7 @@ import { EOL } from "os";
 import { join } from "path";
 import { Disposable, ExtensionContext } from "vscode";
 import { extLogger } from "../common";
-import { readDirFullPaths } from "../pure/files";
+import { readDirFullPaths } from "../common/files";
 import { QueryHistoryDirs } from "./query-history-dirs";
 import { QueryHistoryManager } from "./query-history-manager";
