+++ extensions/ql-vscode/src/variant-analysis/gh-api/mocks/mock-gh-api-server.ts
@@ -6,7 +6,7 @@ import { DisposableObject } from "../../../common/disposable-object";
 
 import { Recorder } from "./recorder";
 import { createRequestHandlers } from "./request-handlers";
-import { getDirectoryNamesInsidePath } from "../../../pure/files";
+import { getDirectoryNamesInsidePath } from "../../../common/files";
 
 /**
  * Enables mocking of the GitHub API server via HTTP interception, using msw.
